Pulled CanvasUITests out into a separate framework under Core
Added some machinery for a test class to be "abstract" meaning that
while it may define tests, they will only be run for the subclasses.

This should allow test reuse between Teacher and Student.
