--- Student/CanvasUITests/Calendar/CalendarTests.swift
@@ -1,49 +1,0 @@
-//
-// This file is part of Canvas.
-// Copyright (C) 2019-present  Instructure, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as
-// published by the Free Software Foundation, either version 3 of the
-// License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-//
-
-import XCTest
-import TestsFoundation
-
-enum Calendar {
-    static var todayButton: Element {
-        return app.find(labelContaining: "Today")
-    }
-
-    static func text(containing text: String) -> Element {
-        return app.find(labelContaining: text)
-    }
-}
-
-class CalendarTests: CanvasUITests {
-    func testCalendarTodayButton() {
-
-        let formatter = DateFormatter()
-        formatter.dateFormat = "MMMM yyyy"
-        let monthYear = formatter.string(from: Date()).uppercased()
-        formatter.dateFormat = "d"
-        let day = formatter.string(from: Date())
-
-        TabBar.calendarTab.tap()
-        app.swipeDown()
-        app.swipeDown()
-        XCTAssertFalse(Calendar.text(containing: monthYear).exists)
-        Calendar.todayButton.tap()
-        Calendar.text(containing: monthYear).waitToExist()
-        Calendar.text(containing: day).waitToExist()
-    }
-}
